Create GoalManager and fixing factories for this change
