+++ DreamRich/goal/models.py
@@ -2,21 +2,19 @@
 All attributes of this file have the annual order of magnitude
 """
 from django.db import models
-from financial_planning.models import FinancialPlanning
-
-
-class FinancialIndependence(models.Model):
-    age = models.PositiveSmallIntegerField()
-    duration_of_usufruct = models.PositiveSmallIntegerField()
-    remain_patrimony = models.PositiveIntegerField()
-    financial_planning = models.OneToOneField(
-        FinancialPlanning,
-        on_delete=models.CASCADE,
-    )
 
 
 class GoalType(models.Model):
     name = models.CharField(max_length=100)
+
+
+class GoalManager(models.Model):
+
+    @property
+    def duration_goals(self):
+        duration_goals = self.financialplanning.financial_independence.age
+
+        return duration_goals
 
 
 class Goal(models.Model):
@@ -25,8 +23,8 @@ class Goal(models.Model):
     year_end = models.PositiveSmallIntegerField()
     periodicity = models.PositiveSmallIntegerField()
     value = models.PositiveIntegerField()
-    financial_planning = models.ForeignKey(
-        FinancialPlanning,
+    goal_manager = models.ForeignKey(
+            GoalManager,
         on_delete=models.CASCADE
     )
     goal_type = models.ForeignKey(GoalType, on_delete=models.CASCADE)
